refactor(agent-tool-context-discovery): rename extract to build and abstract AgentUtilityRegistry
