+++ backend/agentic/agent-tool-context-discovery/src/main/java/org/finos/fluxnova/bpm/engine/ai/agent/discovery/extract/AgentContextSpecBuilder.java
@@ -3,6 +3,6 @@ package org.finos.fluxnova.bpm.engine.ai.agent.discovery.extract;
 import org.finos.fluxnova.bpm.engine.ai.agent.discovery.model.AgentContextSpec;
 import org.finos.fluxnova.bpm.engine.impl.util.xml.Element;
 
-public interface AgentContextSpecExtractor {
-    AgentContextSpec extract(Element adHocElement, String processDefinitionId);
+public interface AgentContextSpecBuilder extends AgentUtilityBuilder<AgentContextSpec> {
+    AgentContextSpec build(Element scopeElement, String processDefinitionId);
 }
