+++ backend/agentic/agent-tool-context-discovery/src/main/java/org/finos/fluxnova/bpm/engine/ai/agent/discovery/extract/AgentToolCatalogueBuilder.java
@@ -3,6 +3,6 @@ package org.finos.fluxnova.bpm.engine.ai.agent.discovery.extract;
 import org.finos.fluxnova.bpm.engine.ai.agent.discovery.model.AgentToolCatalogue;
 import org.finos.fluxnova.bpm.engine.impl.util.xml.Element;
 
-public interface AgentToolCatalogueBuilder {
+public interface AgentToolCatalogueBuilder extends AgentUtilityBuilder<AgentToolCatalogue> {
     AgentToolCatalogue build(Element scopeElement, String processDefinitionId);
 }
