+++ backend/agentic/agent-tool-context-discovery/src/main/java/org/finos/fluxnova/bpm/engine/ai/agent/discovery/extract/AgentUtilityBuilder.java
@@ -1,0 +1,7 @@
+package org.finos.fluxnova.bpm.engine.ai.agent.discovery.extract;
+
+import org.finos.fluxnova.bpm.engine.impl.util.xml.Element;
+
+public interface AgentUtilityBuilder<T> {
+    T build(Element scopeElement, String processDefinitionId);
+}
